[python] support the new CLI system
Just like when the new CLI system was added to iosapplist, this
breaks compatibility with the old CLI syntax.  Particularly, robot
output is now accomplished with `--robot=<format>`, and requesting
apps always accepts either bundle IDs or UUIDs; there is no need to
specify which, and the ability to do so has been removed.