[gui] Add a CLI proxy client for use in the iOS Simulator
The regular CLI can't be used in the simulator because its dyld
will complain that bash (for the appbackup-cli launcher script)
(or even just Python if we weren't using the launcher script) is
compiled for OS X and "can't be used in the simulator" (which is
bullshit), so instead I made a simple proxy that opens a TCP
connection to the `iosappbackup shell-server` running on the OS X
side and provides half-duplex I/O (*).  The Xcode project sets a
preprocessor directive (not committed here) USE_CLI_PROXY to
denote that we're using the proxy.  The directive is set regardless
of whether it's building for the simulator.

The CLI proxy client ignores its command-line flags, so these must
be given when starting the CLI shell server:

    $ iosappbackup \
       -r ~/Library/Developer/CoreSimulator/Devices/<uuid>/data \
       --robot=plist shell-server --null

It listens on 127.0.0.1 (or ::1 with the -6 option) and port 14121
(or the value of --port).  (14121 is chosen because Maserati, the
codename of AppBackup 3.0, was founded on December 1, 1914.)

* The server is full-duplex; however, I'm lazy and I don't want to
  muck with a select(2) loop when the GUI's usage of the CLI is
  half-duplex anyway.